+++ src/main/java/com/cpi/is/controller/maintenance/DispatchTypeController.java
@@ -12,6 +12,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import com.cpi.is.service.maintenance.DispatchTypeService;
+import com.cpi.is.util.SessionUtil;
 
 /**
  * Servlet implementation class DispatchTypeController
@@ -39,19 +40,24 @@ public class DispatchTypeController extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try {
 			action = request.getParameter("action");
-			
-			if ("showDispatchType".equals(action)) {
-				request.setAttribute("dispatchType", new JSONArray(dispatchTypeService.getDispatchType()));
-				page = "pages/navbar/maintenance/dispatchType.jsp";
-			} else if ("saveItem".equals(action)) {
-				String message = dispatchTypeService.saveItem(request);
-				request.setAttribute("message", message);
-				page = "pages/message/success.jsp";
-			} else if ("deleteItem".equals(action)) {
-				String message = dispatchTypeService.deleteItem(request);
-				request.setAttribute("message", message);
-				page = "pages/message/success.jsp";
-			} 
+			if (SessionUtil.isUserLoggedIn(request)) {
+				if ("showDispatchType".equals(action)) {
+					request.setAttribute("dispatchType", new JSONArray(dispatchTypeService.getDispatchType()));
+					page = "pages/navbar/maintenance/dispatchType.jsp";
+				} else if ("saveItem".equals(action)) {
+					String message = dispatchTypeService.saveItem(request);
+					request.setAttribute("message", message);
+					page = "pages/message/success.jsp";
+				} else if ("deleteItem".equals(action)) {
+					String message = dispatchTypeService.deleteItem(request);
+					request.setAttribute("message", message);
+					page = "pages/message/success.jsp";
+				} 
+			} else {
+				page = "/UserController";
+				request.setAttribute("action", "timeout");
+				System.out.println("request is " + request.getAttribute(action));
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 		} finally {
